[SPARK-51252][SS] Add instance metrics for last uploaded snapshot version in HDFS State Stores
### What changes were proposed in this pull request?

SPARK-51252

Similar to SPARK-51097, This PR sets up instance-specific metrics (`SnapshotLastUploaded.partition_<partition id>_<state store name>` to be precise) in the executor side and publishes them through StreamingQueryProgress. The main difference is that this PR adds the metrics to HDFS state stores to have similar observability to RocksDB state stores.

There are two non-test differences compared to the RocksDB PR:

- Add tracking of lastUploadedSnapshotVersion to HDFS state stores, which is updated in the state store object instead of RocksDB. Also updates at the end of writeUpdates instead of uploadSnapshot.
- Metric creation is done in the state store provider class instead of object

The fixes for SparkUI does not affect this PR, since that was an issue with the underlying instance metrics collection process. I still re-verified that these metrics do not show up on SparkUI (screenshot below)

As for the test changes:

- Modified the RocksDB tests slightly to process some additional data in the queries, because the version difference check used in HDFS and RocksDB state stores uses > and >= respectively. The general structure of the tests remain the same.

### Why are the changes needed?

We would like to introduce a similar level of observability into other types of state stores. As for the metrics themselves, it would be helpful in identifying performance degradation issues behind maintenance tasks and more as described in SPARK-51097.

### Does this PR introduce _any_ user-facing change?

There will be some new metrics displayed from StreamingQueryProgress:
```
Streaming query made progress: {
  ...
  "stateOperators" : [ {
    ...
    "customMetrics" : {
      ...
      "SnapshotLastUploaded.partition_0_default" : 2,
      "SnapshotLastUploaded.partition_12_default" : 10,
      "SnapshotLastUploaded.partition_8_default" : 10,
      ...
    }
  } ],
  "sources" : ...,
  "sink" : ...
}
```
Similar to its RocksDB variant, the amount of metrics reported is also tuned using `STATE_STORE_INSTANCE_METRICS_REPORT_LIMIT`, and will by default report 5 instance metrics.

### How was this patch tested?

Four new tests are added into a new suite StateStoreInstanceMetricSuite, making eight tests in total for both state stores. Future instance metrics can be included into this suite for testing and verification.

The first two tests execute a dedup streaming query and verifies metrics are properly filtered + updated through the StreamingQueryProgress logs, but with different StateStore providers that skip maintenance tasks for specific partitions.

The other two tests execute a join streaming query, which contains four state stores per partition instead of one. These two tests verifies metrics are properly collected and filtered as well.

Verified the metrics do not show up on Spark UI:
![image](https://github.com/user-attachments/assets/c3c722fe-3bca-44a3-8f8d-85ed865def05)

### Was this patch authored or co-authored using generative AI tooling?

No.

Closes #50030 from zecookiez/SPARK-51252-hdfs.

Lead-authored-by: Zeyu Chen <zycm03@gmail.com>
